+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## unreleased
 
-- Added: (WOM-344, WOM-367, WOM-457, WOM-466)
+- Added: (PERENTIE-2592, WOM-344, WOM-367, WOM-372, WOM-457, WOM-466)
+  - `TestModeOverrideMixin` class that adds a `test_mode_overrides` attribute and `@overridable` decorator, with logic to apply overrides to Tango attributes when `TestMode` is active. 
   - Copied `Abort` LRC from `SKASubArray` to `SKABaseDevice` that has a command ID and `TaskStatus`.
-  - New user facing LRC attributes `lrcQueue`, `lrcExecuting` and `lrcFinished`. (WOM-372) 
+  - New user facing LRC attributes `lrcQueue`, `lrcExecuting` and `lrcFinished`. () 
   - New private `_lrcEvents` attribute that is meant to be only used by `invoke_lrc`.
   - Added `lrcProtocolVersions` attribute indicating the range of versions of the LRC client-server protocol supported by `SKABaseDevice`. 
     - It is read in `invoke_lrc` to check compatibility between the client and server and to decide which version to use.